break Kinesis tests into appender-only and writer-only
