+++ appenders/src/test/resources/TestKinesisAppender/testStaticClientFactory.properties
@@ -1,14 +1,11 @@
-# config for mock-object testing; note low batch size and delay
+# config for client factory test
 
 log4j.rootLogger=DEBUG, default
 
 log4j.appender.default=com.kdgregory.log4j.testhelpers.aws.kinesis.TestableKinesisAppender
 log4j.appender.default.layout=org.apache.log4j.PatternLayout
-log4j.appender.default.layout.ConversionPattern=%m%n
 
 log4j.appender.default.streamName=argle
 log4j.appender.default.partitionKey=bargle
-log4j.appender.default.batchSize=20
-log4j.appender.default.batchDelay=100
 
-log4j.appender.default.clientFactory=com.kdgregory.log4j.aws.TestKinesisAppender.createMockClient
+log4j.appender.default.clientFactory=com.kdgregory.log4j.aws.TestKinesisLogWriter.createMockClient
